restic: move test helpers and drop unused JSONUnpackedLoader
